Move SchoolList and SchoolDetails out of SchoolsToolPage
