+++ src/pages/tools/schools-tool.js
@@ -18,7 +18,6 @@
 import React, { PropTypes } from 'react';
 import { connect } from 'react-redux';
 import Helmet from 'react-helmet';
-import { Link } from 'react-router';
 import { replace } from 'react-router-redux';
 
 import { uuid4, Immutable as ImmutablePropType } from '../../prop-types/common';
@@ -30,78 +29,12 @@ import { setSchoolsAlphabet } from '../../actions/tools';
 import { ActionsTrigger } from '../../triggers';
 import ApiClient from '../../api/client';
 import { API_HOST } from '../../config';
-import { TAG_SCHOOL } from '../../consts/tags';
 import Button from '../../components/button';
 import VisibilitySensor from '../../components/visibility-sensor';
-import TagIcon from '../../components/tag-icon';
-import FollowTagButton from '../../components/follow-tag-button';
 import AlphabetFilter from '../../components/tools/alphabet-filter';
+import SchoolList from '../../components/tools/school-list';
+import SchoolDetails from '../../components/tools/school-details';
 
-
-function SchoolList({ onClick, schools, selectedSchoolId }) {
-  const items = schools.map((school, index) => {
-    const handleClick = () => onClick(school.get('id'));
-    let className = 'tools_item tools_item-clickable';
-    if (school.get('id') === selectedSchoolId) {
-      className += ' tools_item-selected';
-    }
-
-    return (
-      <div
-        className={className}
-        key={index}
-        onClick={handleClick}
-      >
-        <TagIcon type={TAG_SCHOOL} />
-        <span className="tools_item__child-padded">{school.get('name')}</span>
-        <span
-          className="schools_tool__post_count"
-          title="Number of times this school was used"
-        >
-          ({school.get('post_count')})
-        </span>
-      </div>
-    );
-  });
-
-  return (
-    <div>
-      {items}
-    </div>
-  );
-}
-
-function SchoolDescription({ current_user, school, triggers }) {
-  if (school) {
-    return (
-      <div className="tools_page__details_col">
-        <div className="tools_details">
-          <div className="tools_details__left_col">
-            <span className="micon schools_tool__details_icon">school</span>
-          </div>
-          <div>
-            <Link className="tools_details__title" to={`/s/${school.get('url_name')}`}>
-              {school.get('name')}
-            </Link>
-            <div className="tools_details__description">
-              {school.get('description')}
-            </div>
-            <FollowTagButton
-              className="button-midi"
-              current_user={current_user.toJS()}
-              followed_tags={current_user.get('followed_schools').toJS()}
-              key="follow"
-              tag={school.get('url_name')}
-              triggers={triggers}
-            />
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  return null;
-}
 
 const LIMIT = 25;
 
@@ -260,7 +193,7 @@ class SchoolsToolPage extends React.Component {
             }
           </div>
         </div>
-        <SchoolDescription
+        <SchoolDetails
           current_user={current_user}
           school={selectedSchool}
           triggers={followSchoolTriggers}
